game works now
- took out all of my classes which are now obsolete because all they held was a multiplier, which i can just set at the time of creation anyway
- changed float score to int. makes it easier to see which items you picked up because now it's all using primes. awww yeah, primes.

+++ GameOverScene.h
@@ -10,6 +10,6 @@
 
 @interface GameOverScene : CCScene
 
-@property (nonatomic,assign) float finalScore;
+@property (nonatomic,assign) int finalScore;
 
 @end
